ELT Revision
I am back after a long break from designing with progress going ahead in various sectors of the project. This commitment contains a large rework of the frame/body system for the ELT. The face is now printed seperately to the rest of the body which allows laser cutters/CNC's to directly cutout the face plate from 3mm material. This allows easier access for laser cutting and other non 3d printed related services. The back shroud still needs to be worked around this but that will be coming in a future revision. Laser engravings on printed parts is still a huge work in progress but is undergoing.